hotfix

+++ Labs_Java/src/Lab_7/ArraySum.java
@@ -10,30 +10,30 @@ public class ArraySum {
     public static void main(String[] args) throws Exception {
         final int ARRAY_SIZE = 100_000_000;
 
-        long[] myarr = new long[ARRAY_SIZE];
+        int[] myarr = new int[ARRAY_SIZE];
 
         for (int i = 0; i < ARRAY_SIZE; i++) {
-            myarr[i] = (long) i * i;
+            myarr[i] = i * i;
         }
 
-        Callable<Long> task1 = createTask(myarr, 1, ARRAY_SIZE);
-        Callable<Long> task2 = createTask(myarr, 2, ARRAY_SIZE);
-        Callable<Long> task3 = createTask(myarr, 3, ARRAY_SIZE);
-        Callable<Long> task4 = createTask(myarr, 4, ARRAY_SIZE);
+        Callable<Integer> task1 = createTask(myarr, 1, ARRAY_SIZE);
+        Callable<Integer> task2 = createTask(myarr, 2, ARRAY_SIZE);
+        Callable<Integer> task3 = createTask(myarr, 3, ARRAY_SIZE);
+        Callable<Integer> task4 = createTask(myarr, 4, ARRAY_SIZE);
 
         ExecutorService executor = Executors.newFixedThreadPool(4);
 
-        Future<Long> maxFirst = executor.submit(task1);
-        Future<Long> maxSecond = executor.submit(task2);
-        Future<Long> maxThird = executor.submit(task3);
-        Future<Long> maxFourth = executor.submit(task4);
+        Future<Integer> maxFirst = executor.submit(task1);
+        Future<Integer> maxSecond = executor.submit(task2);
+        Future<Integer> maxThird = executor.submit(task3);
+        Future<Integer> maxFourth = executor.submit(task4);
 
-        long max1 = maxFirst.get();
-        long max2 = maxSecond.get();
-        long max3 = maxThird.get();
-        long max4 = maxFourth.get();
+        int max1 = maxFirst.get();
+        int max2 = maxSecond.get();
+        int max3 = maxThird.get();
+        int max4 = maxFourth.get();
 
-        long overallMax = Math.max(Math.max(max1, max2), Math.max(max3, max4));
+        int overallMax = Math.max(Math.max(max1, max2), Math.max(max3, max4));
 
         System.out.println("Maximum threaded: " + overallMax);
 
@@ -41,7 +41,7 @@ public class ArraySum {
 
     }
 
-    private static Callable<Long> createTask(long[] array, int part, int arraySize) {
+    private static Callable<Integer> createTask(int[] array, int part, int arraySize) {
         return () -> {
             int start, end;
 
@@ -64,7 +64,7 @@ public class ArraySum {
                     break;
             }
 
-            long maxi = 0;
+            int maxi = 0;
             for (int i = start; i < end; i++) {
                 if (maxi < array[i]) maxi = array[i];
             }
